Former-commit-id: f53d7fa8cf11ed77f02ad435e399c134f867d694

+++ acrg_hbmcmc/hbmcmc.py
@@ -20,17 +20,24 @@ being annoying it will also slow down your run due to unnecessary forking.
 
 """
 import os
+import sys
 import numpy as np
 import acrg_name as name
 import acrg_obs as getobs
 import socket
 import acrg_hbmcmc.inversionsetup as setup 
 import acrg_hbmcmc.inversion_pymc3 as mcmc
-import acrg_hbmcmc.quadtreebasis as quadtree
+# import acrg_hbmcmc.quadtreebasis as quadtree
+import acrg_name.basis_functions as basis
 import shutil
 
-acrg_path = os.getenv("ACRG_PATH")
-data_path = os.getenv("DATA_PATH")
+if sys.version_info[0] == 2: # If major python version is 2, can't use paths module
+    acrg_path = os.getenv("ACRG_PATH")
+    data_path = os.getenv("DATA_PATH") 
+else:
+    from acrg_config.paths import paths
+    acrg_path = paths.acrg
+    data_path = paths.data
 
 def fixedbasisMCMC(species, sites, domain, meas_period, start_date, 
                    end_date, outputpath, outputname, 
@@ -181,7 +188,7 @@ def fixedbasisMCMC(species, sites, domain, meas_period, start_date,
             print("Basis case %s supplied but quadtree_basis set to True" % fp_basis_case)
             print("Assuming you want to use %s " % fp_basis_case)
         else:
-            tempdir = quadtree.quadtreebasisfunction(emissions_name, fp_all, sites, 
+            tempdir = basis.quadtreebasisfunction(emissions_name, fp_all, sites, 
                           start_date, domain, species, outputname,
                           nbasis=nbasis)
             fp_basis_case= "quadtree"+species+"-"+outputname
@@ -241,9 +248,10 @@ def fixedbasisMCMC(species, sites, domain, meas_period, start_date,
                            emissions_name, domain, species, sites,
                            start_date, end_date, outputname, outputpath,
                            basis_directory, country_file, fp_basis_case, country_unit_prefix)
-    
-    # remove the temporary basis function directory
-    shutil.rmtree(tempdir)
+
+    if quadtree_basis is True:
+        # remove the temporary basis function directory
+        shutil.rmtree(tempdir)
     
     print("All done")
